make mask wide enough

+++ src/devices/bus/nes/act53.cpp
@@ -111,7 +111,7 @@ void nes_action53_device::update_prg()
 {
 	u16 prg_lo, prg_hi;
 	u8 size = (m_reg[2] & 0x30) >> 4;         // Game size
-	u8 mask = ~0 << (size + 1);               // Bits to be taken from PRG regs
+	u16 mask = ~0 << (size + 1);              // Bits to be taken from PRG regs
 	u8 b32k = !BIT(m_reg[2], 3);              // 32K mode bit
 	u16 outer = m_reg[3] << 1;                // Outer PRG reg bits
 	u8 inner = (m_reg[1] << b32k) & ~mask;    // Inner PRG reg bits
